# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PSOTS Community Portal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: notices
  - name: events
  - name: listings
  - name: contacts

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List active notices
      parameters:
        - name: tower
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: archived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
    post:
      operationId: createNotice
      tags: [notices]
      summary: Create a notice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeInput"
      responses:
        "201":
          description: Created notice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"

  /notices/{id}:
    get:
      operationId: getNotice
      tags: [notices]
      summary: Get a notice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List marketplace listings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a marketplace listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"

  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingInput"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List emergency contacts
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Notice:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        tower:
          type: string
        isPinned:
          type: boolean
        postedBy:
          type: string
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, content, category, tower, isPinned, postedBy, createdAt]

    CreateNoticeInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        tower:
          type: string
        isPinned:
          type: boolean
        postedBy:
          type: string
      required: [title, content, category, tower, postedBy]

    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        location:
          type: string
        eventDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        organizer:
          type: string
        rsvpCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, location, eventDate, organizer, rsvpCount, createdAt]

    CreateEventInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        eventDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        organizer:
          type: string
      required: [title, description, location, eventDate, organizer]

    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
          nullable: true
        type:
          type: string
        category:
          type: string
        tower:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
          nullable: true
        isActive:
          type: boolean
        status:
          type: string
        telegramUserId:
          type: string
          nullable: true
        telegramUsername:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, type, category, tower, contactName, isActive, status, createdAt]

    CreateListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
          nullable: true
        type:
          type: string
        category:
          type: string
        tower:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
          nullable: true
        telegramUserId:
          type: string
          nullable: true
        telegramUsername:
          type: string
          nullable: true
      required: [title, description, type, category, tower, contactName]

    UpdateListingInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, sold, closed]
      required: [status]

    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        available:
          type: string
      required: [id, name, role, phone, available]