# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSOTS Community Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  role: string;
  phone: string;
  available: string;
}