# lib/api-zod/src/generated/types/createEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSOTS Community Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventInput {
  title: string;
  description: string;
  location: string;
  eventDate: Date;
  endDate?: Date | null;
  organizer: string;
}