# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSOTS Community Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  title: string;
  description: string;
  price?: number | null;
  type: string;
  category: string;
  tower: string;
  contactName: string;
  contactPhone?: string | null;
  isActive: boolean;
  status: string;
  telegramUserId?: string | null;
  telegramUsername?: string | null;
  expiresAt?: Date | null;
  createdAt: Date;
}