# lib/api-zod/src/generated/types/notice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSOTS Community Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Notice {
  id: number;
  title: string;
  content: string;
  category: string;
  tower: string;
  isPinned: boolean;
  postedBy: string;
  archivedAt?: Date | null;
  createdAt: Date;
}