# lib/api-zod/src/generated/types/updateListingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSOTS Community Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateListingInputStatus } from "./updateListingInputStatus";

export interface UpdateListingInput {
  status: UpdateListingInputStatus;
}